Use Guava's lerp and clamp; minor cleanups

+++ core/src/main/java/com/radixdlt/consensus/liveness/Pacemaker.java
@@ -466,7 +466,7 @@ public final class Pacemaker implements BFTEventProcessorAtCurrentRound {
   private void rescheduleTimeout(ScheduledLocalTimeout scheduledTimeout) {
     final var timeout =
         timeoutCalculator.calculateTimeoutMs(
-            scheduledTimeout.count() + 1, vertexStore.getCurrentUtilizationRatio());
+            (long) scheduledTimeout.count() + 1, vertexStore.getCurrentUtilizationRatio());
     final var nextTimeout = scheduledTimeout.nextRetry(timeout);
     this.scheduledLocalTimeoutDispatcher.dispatch(nextTimeout, timeout);
   }
